backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pandas
requests
python-dotenv
